Fix specs with a more conservative approach
