+++ core/app/models/spree/behaviors/advance_shipments.rb
@@ -2,7 +2,6 @@ module Spree
   module Behaviors
     class AdvanceShipments < OrderBase
       def run
-        return unless order.completed?
         order.shipments.select(&:persisted?).each { |shipment| shipment.update!(order) }
       end
     end
