--- core/app/models/spree/behaviors/persist_order.rb
@@ -1,11 +1,0 @@
-module Spree
-  module Behaviors
-    class PersistOrder < OrderBase
-
-      def run
-        order.save!
-      end
-
-    end
-  end
-end
